# Add simpler variants of visit for perf