## Before
Grats completed in: 8.246s
Grats completed in: 7.723s
Grats completed in: 7.734s
Grats completed in: 7.871s

## After
Grats completed in: 7.373s
Grats completed in: 7.307s
Grats completed in: 7.546s
Grats completed in: 7.444s